+++ optimus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.4"
+__version__ = "0.12.5"
 
 
 def safe_commit() -> None: